+++ api/src/Blog/Create.ts
@@ -1,8 +1,8 @@
 import { UserRepo } from "#Accounts/UserRepo"
 import { handlerFor } from "#api/lib/handler"
 import { OperationsDefault } from "#api/lib/layers"
-import { S } from "#api/resources/lib"
 import { Events } from "#api/services"
+import { S } from "#resources/lib"
 import { Effect } from "effect-app"
 import { InvalidStateError, NotFoundError, OptimisticConcurrencyException } from "effect-app/client"
 import { BlogPost, BlogPostId } from "./models.js"
