+++ api/src/Blog/List.ts
@@ -1,5 +1,5 @@
 import { handlerFor } from "#api/lib/handler"
-import { S } from "#api/resources/lib"
+import { S } from "#resources/lib"
 import { Effect } from "effect-app"
 import { BlogPostView } from "./PostView.js"
 import { BlogPostRepo } from "./Repo.js"
