+++ api/src/Blog/Publish.ts
@@ -1,7 +1,7 @@
 import { handlerFor } from "#api/lib/handler"
 import { OperationsDefault } from "#api/lib/layers"
-import { BogusEvent } from "#api/resources/Events"
-import { S } from "#api/resources/lib"
+import { BogusEvent } from "#resources/Events"
+import { S } from "#resources/lib"
 import { Events, Operations } from "#api/services"
 import { Duration, Effect, Schedule } from "effect-app"
 import { NotFoundError } from "effect-app/client"
